Move Counts and prob_of_alignments to base module because it is shared by Models 4 and 5
Also remove Counts.null and compute it on demand in
update_null_generation
